+++ src/main/java/org/jsoup/nodes/Element.java
@@ -1063,11 +1063,13 @@ public class Element extends Node implements Iterable<Element> {
      */
     @Override
     public Element empty() {
-        // Detach each of the children -> parent links:
-        int size = childNodes.size();
-        for (int i = 0; i < size; i++)
-            childNodes.get(i).parentNode = null;
-        childNodes.clear();
+        if (childNodes != EmptyNodeList) {
+            // Detach each of the children -> parent links:
+            int size = childNodes.size();
+            for (int i = 0; i < size; i++)
+                childNodes.get(i).parentNode = null;
+            childNodes.clear();
+        }
         return this;
     }